Revert "Update modal form data and fix RR_BASE error"
