--- ITEM RARITY/README_MODULARIZATION.md
@@ -1,103 +1,0 @@
-# ITEM RARITY System Modularization
-
-## Overview
-The ITEM RARITY system has been modularized by separating stats, skills, and passives triggers into separate script files for better organization and maintainability.
-
-## New File Structure
-
-### 📁 stats.js
-Contains all stats-related functionality:
-- `getStatsFromLore(itemStack)` - Extracts stats from item lore
-- `statsMainMenu(player)` - Main stats menu interface
-- `showEnhancedStatsRerollMenu(equipment, player, itemStack)` - Stats reroll interface
-- `showEnhancedStatsUpgradeMenu(equipment, player, itemStack)` - Stats upgrade interface
-- `performBulkStatsReroll(equipment, player, itemStack, adjustedCost, attempts, startingCounter, lockedStats)` - Bulk stats rerolling
-- `performBulkStatsUpgrade(equipment, player, itemStack, adjustedCost, upgradeCost, attempts, startingCounter)` - Bulk stats upgrading
-
-### 📁 skills.js
-Contains all skills-related functionality:
-- `showEnhancedSkillUpgradeMenu(equipment, player, itemStack)` - Skills upgrade interface
-- `showEnhancedSkillRerollMenu(equipment, player, itemStack)` - Skills reroll interface
-- `performBulkSkillUpgrade(equipment, player, itemStack, adjustedCost, upgradeCost, attempts, startingCounter)` - Bulk skills upgrading
-- `performBulkSkillReroll(equipment, player, itemStack, adjustedCost, attempts, startingCounter)` - Bulk skills rerolling
-- Helper functions for skill manipulation (placeholders for implementation)
-
-### 📁 passives.js
-Contains all passives-related functionality:
-- `showEnhancedPassiveUpgradeMenu(equipment, player, itemStack)` - Passives upgrade interface
-- `showEnhancedPassiveRerollMenu(equipment, player, itemStack)` - Passives reroll interface
-- `performBulkPassiveUpgrade(equipment, player, itemStack, adjustedCost, upgradeCost, attempts, startingCounter)` - Bulk passives upgrading
-- `performBulkPassiveReroll(equipment, player, itemStack, adjustedCost, attempts, startingCounter)` - Bulk passives rerolling
-- Helper functions for passive manipulation (placeholders for implementation)
-
-### 📁 main.js (Updated)
-- ✅ Updated imports to include new modular functions
-- ✅ Added exports for utility functions used by separate scripts
-- ✅ Updated to use 2.1.0 beta ModalFormData API
-- ✅ Fixed RR_BASE undefined error
-
-## Changes Made
-
-### ✅ Completed
-1. **API Migration**: All ModalFormData calls updated to 2.1.0 beta format
-2. **New Modules Created**: stats.js, skills.js, passives.js with proper exports
-3. **Import Structure**: Added imports in main.js for modular functions
-4. **Utility Exports**: Exported required utility functions from main.js
-5. **Cross-file Dependencies**: Set up proper import/export relationships
-
-### ⚠️ Still Required
-The following functions need to be **removed from main.js** since they're now in separate files:
-
-**Stats Functions to Remove:**
-- `showEnhancedStatsRerollMenu()` (around line 1623)
-- `showEnhancedStatsUpgradeMenu()` (around line 1687)
-- `performBulkStatsReroll()` (around line 2398)
-- `performBulkStatsUpgrade()` (around line 2477)
-
-**Skills Functions to Remove:**
-- `showEnhancedSkillUpgradeMenu()` (around line 1736)
-- `showEnhancedSkillRerollMenu()` (around line 1780)
-- `performBulkSkillUpgrade()` (around line 2540)
-- `performBulkSkillReroll()` (around line 2570)
-
-**Passives Functions to Remove:**
-- `showEnhancedPassiveUpgradeMenu()` (around line 1824)
-- `showEnhancedPassiveRerollMenu()` (around line 1868)
-- `performBulkPassiveUpgrade()` (around line 2595)
-- `performBulkPassiveReroll()` (around line 2625)
-
-### 🔧 Additional Helper Functions Needed
-The separate files contain placeholder functions that need implementation:
-
-**stats.js:**
-- `generateRandomStat(rarity)` - Generate random stat based on rarity
-- `updateItemStatsInLore(itemStack, newStats)` - Update item lore with new stats
-- `upgradeStatValue(stat, multiplier)` - Upgrade a stat value
-
-**skills.js:**
-- `upgradeItemSkills(itemStack, multiplier)` - Upgrade item skills
-- `rerollItemSkills(itemStack, rarity)` - Reroll item skills
-
-**passives.js:**
-- `upgradeItemPassives(itemStack, multiplier)` - Upgrade item passives
-- `rerollItemPassives(itemStack, rarity)` - Reroll item passives
-
-## Benefits of Modularization
-
-1. **Better Organization**: Related functionality grouped together
-2. **Easier Maintenance**: Changes to stats/skills/passives are isolated
-3. **Reduced File Size**: main.js is now more manageable
-4. **Clear Dependencies**: Import/export structure shows relationships
-5. **Reusability**: Functions can be easily imported by other modules
-6. **Testing**: Individual modules can be tested independently
-
-## Usage
-
-The modular system maintains the same external API. All functions work exactly as before, but are now properly organized across multiple files. The import/export system ensures all dependencies are correctly resolved.
-
-## Next Steps
-
-1. Remove duplicate functions from main.js (listed above)
-2. Implement the placeholder helper functions
-3. Test all functionality to ensure imports work correctly
-4. Consider further modularization if beneficial (e.g., separate UI components)
